Refactor installer layout into modular packages
