+++ internal/kumi/fs.go
@@ -1,41 +1,42 @@
-package fsutil
+package kumi
 
 import (
 	"os"
 	"path/filepath"
+	"strings"
 )
 
-func EnsureDir(path string) error {
+func ensureDir(path string) error {
 	return os.MkdirAll(path, 0o755)
 }
 
-func PathExists(path string) bool {
-	if path == "" {
+func pathExists(path string) bool {
+	if strings.TrimSpace(path) == "" {
 		return false
 	}
 	_, err := os.Stat(path)
 	return err == nil
 }
 
-func FirstExisting(candidates []string, exeName string) string {
+func firstExisting(candidates []string, exeName string) string {
 	for _, candidate := range candidates {
-		if candidate == "" {
+		if strings.TrimSpace(candidate) == "" {
 			continue
 		}
 		probe := candidate
 		if filepath.Ext(candidate) == "" {
 			probe = filepath.Join(candidate, exeName)
 		}
-		if PathExists(probe) {
+		if pathExists(probe) {
 			return probe
 		}
 	}
 	return ""
 }
 
-func FirstExistingDirectory(candidates []string) string {
+func firstExistingDirectory(candidates []string) string {
 	for _, candidate := range candidates {
-		if candidate == "" {
+		if strings.TrimSpace(candidate) == "" {
 			continue
 		}
 		info, err := os.Stat(candidate)
