+++ internal/kumi/install/feather.go
@@ -1,0 +1,7 @@
+package install
+
+import ktypes "polyforge/internal/kumi/types"
+
+func Feather(deps Dependencies, candidates []string, zipURL string, warning string) (*ktypes.ActionResult, error) {
+	return InstanceWithOptionalZip(deps, "Feather", candidates, "profiles", "TurtelSMP5", zipURL, warning)
+}
